requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
openai==1.65.1
markdown==3.5.1
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
gunicorn==21.2.0
anthropic==0.49.0
psycopg2-binary==2.9.9
channels>=4.0.0
daphne==4.0.0
channels-redis==4.1.0
websockets==15.0.1
google-genai==1.9.0
stripe==7.9.0
paramiko==3.3.1
asgiref>=3.6.0
kubernetes>=28.1.0
urllib3>=1.26.0
pyyaml>=6.0
websocket-client>=1.6.0 